backend/requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
colorama==0.4.6
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
packaging==24.1
pathspec==0.12.1
Pillow==10.4.0
platformdirs==4.3.1
sqlparse==0.5.1
tzdata==2024.1
